+++ liboptical/wash_opt.cpp
@@ -111,7 +111,7 @@ void WashCrewOptimizer::ReadFromFiles()
 	{
 		m_solar_data.num_mirrors_by_group[i] = 1;
 	}
-	m_solar_data.num_mirror_groups = hel_id.size();
+	m_solar_data.num_mirror_groups = (int)hel_id.size();
 	
 }
 
@@ -315,7 +315,7 @@ void WashCrewOptimizer::GroupSolutionMirrors(int hours)
 		m_solution_data.num_mirrors_by_group[i] = num_mirrors_by_group.at(i);
 	}
 	m_solution_data.total_mirror_output = m_condensed_data.total_mirror_output*1.0;
-	m_solution_data.num_mirror_groups = mirror_output.size();
+	m_solution_data.num_mirror_groups = (int)mirror_output.size();
 	m_results.assignments = new_assignments;
 }
 
@@ -635,7 +635,7 @@ std::vector<int> WashCrewOptimizer::RetracePath(
 		path.push_back(parent);
 		parent = parents[i*row_length + parent];
 	}
-	for (int i = path.size()-1; i >= 0; i--)
+	for (int i = (int)(path.size()-1); i >= 0; i--)
 	{
 		rev_path.push_back(path.at(i));
 	}
@@ -704,7 +704,7 @@ void WashCrewOptimizer::OptimizeWashCrews(int scale, bool output)
 	std::cerr << "optimal cost: " << min_cost << "\nNumber of wash crews: "
 		<< best_path.size()-1 << "\n";
 	m_results.assignments = best_path;
-	m_results.num_wash_crews = best_path.size() - 1;
+	m_results.num_wash_crews = (int)(best_path.size() - 1);
 	if (output)
 	{
 		OutputResults();
